+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # Changelog
 
 ## 0.22.2
-
-<!-- Prepare for next release: remove this line and write your release notes -->
-
+### Improvements
+- Added new command `configure jira issue defaults` to setup payload for project where to create jira tasks from thread
 
 ## 0.22.1
 ### Bugfixes